refactor(dialog-remote-ucan-s3): factor the cached-redeem block into one helper
The nine-line cached-redeem-then-invoke block was duplicated verbatim
across all seven ForkInvocation<UcanSite, _> provider impls. Factor it
into permit_cache::execute_cached, generic over the effect type, and
collapse each impl to a single call. Return types and error-conversion
semantics are unchanged.
