+++ rust/dialog-remote-ucan-s3/src/provider/archive.rs
@@ -4,7 +4,6 @@ use async_trait::async_trait;
 use dialog_capability::ForkInvocation;
 use dialog_capability::Provider;
 use dialog_effects::archive::*;
-use dialog_remote_s3::S3;
 
 use crate::site::UcanSite;
 
@@ -15,19 +14,7 @@ impl Provider<ForkInvocation<UcanSite, Get>> for UcanSite {
         &self,
         invocation: ForkInvocation<UcanSite, Get>,
     ) -> Result<Option<Vec<u8>>, ArchiveError> {
-        let (permit, key) = crate::permit_cache::redeem_cached(
-            &invocation.authorization,
-            &invocation.address,
-            &invocation.capability,
-        )
-        .await?;
-        let result = permit.invoke(invocation.capability).perform(&S3).await;
-        if result.is_err() {
-            // A permit that failed downstream may be stale (revoked or
-            // expired server-side); drop it so the next attempt redeems.
-            crate::permit_cache::PermitCache::shared().invalidate(&key);
-        }
-        result
+        crate::permit_cache::execute_cached(invocation).await
     }
 }
 
@@ -35,18 +22,6 @@ impl Provider<ForkInvocation<UcanSite, Get>> for UcanSite {
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 impl Provider<ForkInvocation<UcanSite, Put>> for UcanSite {
     async fn execute(&self, invocation: ForkInvocation<UcanSite, Put>) -> Result<(), ArchiveError> {
-        let (permit, key) = crate::permit_cache::redeem_cached(
-            &invocation.authorization,
-            &invocation.address,
-            &invocation.capability,
-        )
-        .await?;
-        let result = permit.invoke(invocation.capability).perform(&S3).await;
-        if result.is_err() {
-            // A permit that failed downstream may be stale (revoked or
-            // expired server-side); drop it so the next attempt redeems.
-            crate::permit_cache::PermitCache::shared().invalidate(&key);
-        }
-        result
+        crate::permit_cache::execute_cached(invocation).await
     }
 }
