+++ rust/dialog-remote-ucan-s3/src/provider/blob.rs
@@ -10,7 +10,6 @@ use async_trait::async_trait;
 use dialog_capability::ForkInvocation;
 use dialog_capability::Provider;
 use dialog_effects::blob::{BlobError, BlobReader, BlobWriter, Import, Read};
-use dialog_remote_s3::S3;
 
 use crate::site::UcanSite;
 
@@ -21,19 +20,7 @@ impl Provider<ForkInvocation<UcanSite, Read>> for UcanSite {
         &self,
         invocation: ForkInvocation<UcanSite, Read>,
     ) -> Result<BlobReader, BlobError> {
-        let (permit, key) = crate::permit_cache::redeem_cached(
-            &invocation.authorization,
-            &invocation.address,
-            &invocation.capability,
-        )
-        .await?;
-        let result = permit.invoke(invocation.capability).perform(&S3).await;
-        if result.is_err() {
-            // A permit that failed downstream may be stale (revoked or
-            // expired server-side); drop it so the next attempt redeems.
-            crate::permit_cache::PermitCache::shared().invalidate(&key);
-        }
-        result
+        crate::permit_cache::execute_cached(invocation).await
     }
 }
 
@@ -44,18 +31,6 @@ impl Provider<ForkInvocation<UcanSite, Import>> for UcanSite {
         &self,
         invocation: ForkInvocation<UcanSite, Import>,
     ) -> Result<BlobWriter, BlobError> {
-        let (permit, key) = crate::permit_cache::redeem_cached(
-            &invocation.authorization,
-            &invocation.address,
-            &invocation.capability,
-        )
-        .await?;
-        let result = permit.invoke(invocation.capability).perform(&S3).await;
-        if result.is_err() {
-            // A permit that failed downstream may be stale (revoked or
-            // expired server-side); drop it so the next attempt redeems.
-            crate::permit_cache::PermitCache::shared().invalidate(&key);
-        }
-        result
+        crate::permit_cache::execute_cached(invocation).await
     }
 }
